Patch v1.0.1
Fixed recursion errors caused by Endpoint constructors as well as
casting errors. Changed ID format from Integer to Long.

+++ bolt-java-wrapper/src/main/java/watch/bolt/api/core/game/Game.java
@@ -4,7 +4,7 @@ import watch.bolt.api.core.BoltObject;
 
 public class Game implements BoltObject {
 
-	protected int id;
+	protected long id;
 	protected String name;
 
 	public Game(int id, String name) {
@@ -12,7 +12,7 @@ public class Game implements BoltObject {
 		this.name = name;
 	}
 
-	public int getID() {
+	public long getID() {
 		return id;
 	}
